*Update rAthena 56ac0eeae6, 9e50599234, 1ebe1ada8a, 8ad5d4f7c4, e0397a6e34

+++ trunk/src/map/homunculus.h
@@ -56,7 +56,7 @@ struct homun_data {
 	struct s_homunculus_db *homunculusDB; //[orn]
 	struct s_homunculus homunculus; //[orn]
 
-	int64 masterteleport_timer;
+	int masterteleport_timer;
 	struct map_session_data *master; //Pointer back to its master
 	int hungry_timer; //[orn]
 	unsigned int exp_next;
